Updated copyright year

+++ Esp32Mock/PubSubClient.cpp
@@ -1,4 +1,4 @@
-// Copyright 2021-2022 Rik Essenius
+// Copyright 2021-2023 Rik Essenius
 // 
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
 // except in compliance with the License. You may obtain a copy of the License at
@@ -48,10 +48,10 @@ bool PubSubClient::loop() {
 }
 
 bool PubSubClient::publish(const char* topic, const char* payload, bool retain) {
-    if (strlen(topic) + strlen(_topic) > TopicSize - 1) return false;
+    if (strlen(topic) + strlen(_topic) > static_cast<size_t>(TopicSize) - 1) return false;
     SafeCString::strcat(_topic, topic);
     SafeCString::strcat(_topic, "\n");
-    if (strlen(payload) + strlen(_payload) > PayloadSize - 1) return false;
+    if (strlen(payload) + strlen(_payload) > static_cast<size_t>(PayloadSize) - 1) return false;
     SafeCString::strcat(_payload, payload);
     if (!retain) SafeCString::strcat(_payload, "[x]");
     SafeCString::strcat(_payload, "\n");
